moved step to sharedSteps, changes from pr
